+++ src/components/loaders/internet-loader.jsx
@@ -3,11 +3,11 @@ const Internetloader = () => {
     <>
       <div className="internet-loader-background">
         <div className="internet-loader-box">
-            <span class="internet-loader-spinner"></span>
+            <span className="internet-loader-spinner"></span>
         </div>
       </div>
     </>
   )
 }
 
-export default Internetloader;
+export default Internetloader