+++ src/index.js
@@ -2,11 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 
-import Internetloader from './components/loaders/internetloader';
+import GameLoader from './components/loaders/game-loader';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Internetloader />
+    <GameLoader />
   </React.StrictMode>
 );